backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
PyJWT==2.9.0
sqlparse==0.5.3
